% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_functions.R
\name{qs_deserialize}
\alias{qs_deserialize}
\title{qs_deserialize}
\usage{
qs_deserialize(input, validate_checksum = FALSE, nthreads = 1L)
}
\arguments{
\item{input}{The raw vector to deserialize.}

\item{validate_checksum}{Whether to validate the stored checksum in the file (default \code{FALSE}). This can be used to test for file corruption but has a performance penality.}

\item{nthreads}{The number of threads to use when reading data (default: \code{1}).}
}
\value{
The deserialized object.
}
\description{
Deserializes a raw vector to an object using the \code{qs2} format.
}
\examples{
x <- data.frame(int = sample(1e3, replace=TRUE),
        num = rnorm(1e3),
        char = sample(state.name, 1e3, replace=TRUE),
        stringsAsFactors = FALSE)
xserialized <- qs_serialize(x)
x2 <- qs_deserialize(xserialized)
identical(x, x2) # returns true
}
