% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_functions.R
\name{qx_dump}
\alias{qx_dump}
\title{qx_dump}
\usage{
qx_dump(file)
}
\arguments{
\item{file}{A file name/path.}
}
\value{
The uncompressed serialization.
}
\description{
Exports the uncompressed binary serialization to a list of raw vectors for both \code{qs2} and \code{qdata} formats.
For testing and exploratory purposes mainly.
}
\examples{
x <- data.frame(int = sample(1e3, replace=TRUE),
        num = rnorm(1e3),
        char = sample(state.name, 1e3, replace=TRUE),
        stringsAsFactors = FALSE)
myfile <- tempfile()
qs_save(x, myfile)
binary_data <- qx_dump(myfile)
}
