% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_functions.R
\name{xxhash_raw}
\alias{xxhash_raw}
\title{XXH3_64 hash}
\usage{
xxhash_raw(data)
}
\arguments{
\item{data}{The data to hash.}
}
\value{
The 64-bit hash.
}
\description{
Calculates a 64-bit XXH3 hash.
}
\examples{
x <- as.raw(c(1,2,3))
xxhash_raw(x)
}
