% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{qd_save}
\alias{qd_save}
\title{qd_save}
\usage{
qd_save(object, file, compress_level = 3L,
shuffle = TRUE, warn_unsupported_types=TRUE,
nthreads = 1L)
}
\description{
Saves an object to disk using the \code{qdata} format.
}
\examples{
x <- data.frame(int = sample(1e3, replace=TRUE),
        num = rnorm(1e3),
        char = sample(state.name, 1e3, replace=TRUE),
         stringsAsFactors = FALSE)
myfile <- tempfile()
qd_save(x, myfile)
x2 <- qd_read(myfile)
identical(x, x2) # returns true

# qdata support multithreading
qd_save(x, myfile, nthreads=1)
x2 <- qd_read(myfile, nthreads=1)
identical(x, x2) # returns true
}
