% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{qs_save}
\alias{qs_save}
\title{qs_save}
\usage{
qs_save(object, file, compress_level = 3L,
shuffle = TRUE, nthreads = 1L)
}
\description{
Saves an object to disk using the \code{qs2} format.
}
\examples{
x <- data.frame(int = sample(1e3, replace=TRUE),
        num = rnorm(1e3),
        char = sample(state.name, 1e3, replace=TRUE),
         stringsAsFactors = FALSE)
myfile <- tempfile()
qs_save(x, myfile)
x2 <- qs_read(myfile)
identical(x, x2) # returns true

# qs2 support multithreading
qs_save(x, myfile, nthreads=1)
x2 <- qs_read(myfile, nthreads=1)
identical(x, x2) # returns true
}
