% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_functions.R
\name{zstd_compress_bound}
\alias{zstd_compress_bound}
\title{Zstd compress bound}
\usage{
zstd_compress_bound(size)
}
\arguments{
\item{size}{An integer size}
}
\value{
Maximum compressed size.
}
\description{
Exports the compress bound function from the zstd library. Returns the maximum potential compressed size of an object of length \code{size}.
}
\examples{
zstd_compress_bound(100000)
zstd_compress_bound(1e9)
}
